+++ PurlemonHazel/src/Platform/OpenGL/OpenGLShader.h
@@ -26,7 +26,7 @@ namespace PH {
 		void UploadUniformMat3(const std::string& name, const glm::mat3& matrix);
 		void UploadUniformMat4(const std::string& name, const glm::mat4& matrix);
 		
-		void UploadUniformInt(const std::string& name, float value);
+		void UploadUniformInt(const std::string& name, int value);
 
 		void UploadUniformFloat(const std::string& name, float value);
 		void UploadUniformFloat2(const std::string& name, const glm::vec2& vec);
